requirements.txt
beir
datasets
tqdm
bm25s
sentence-transformers
faiss-cpu
numpy
streamlit